Updates the app deployment for Openshift

+++ bin/test-workflow/6_app_deploy_backend.sh
@@ -17,7 +17,7 @@ set_namespace "$TEST_APP_NAMESPACE_NAME"
 app_name="app-backend-pg"
 
 # Uninstall backend if it exists so any PVCs can be deleted
-if [ "$(helm list -q -n $TEST_APP_NAMESPACE_NAME | grep "^$app_name$")" = "$app_name" ]; then
+if [ "$(helm list -q -a -n $TEST_APP_NAMESPACE_NAME | grep "^$app_name$")" = "$app_name" ]; then
     helm uninstall "$app_name" -n "$TEST_APP_NAMESPACE_NAME"
 fi
 
@@ -38,7 +38,8 @@ $cli --namespace "$TEST_APP_NAMESPACE_NAME" \
 helm repo add bitnami https://charts.bitnami.com/bitnami
 helm repo update
 
-helm install "$app_name" bitnami/postgresql -n "$TEST_APP_NAMESPACE_NAME" --debug --wait --timeout "$TIMEOUT" \
+
+args=(install "$app_name" bitnami/postgresql -n "$TEST_APP_NAMESPACE_NAME" --debug --wait --timeout "$TIMEOUT" \
     --set image.repository="postgres" \
     --set image.tag="9.6" \
     --set postgresqlDataDir="/data/pgdata" \
@@ -52,5 +53,18 @@ helm install "$app_name" bitnami/postgresql -n "$TEST_APP_NAMESPACE_NAME" --debu
     --set securityContext.fsGroup="999" \
     --set postgresqlDatabase="test_app" \
     --set postgresqlUsername="test_app" \
-    --set postgresqlPassword="$SAMPLE_APP_BACKEND_DB_PASSWORD"
+    --set postgresqlPassword="$SAMPLE_APP_BACKEND_DB_PASSWORD")
+
+# For Openshift the runAsUser and fsgroup values are different than Kubernetes,
+# and we opt to use the dynamic settings. See the below link for nore info. 
+# https://github.com/bitnami/charts/tree/master/bitnami/postgresql#differences-between-bitnami-postgresql-image-and-docker-official-image
+if [[ "$PLATFORM" == "openshift" ]]; then
+  args+=(--set containerSecurityContext.enabled=false \
+    --set volumePermissions.securityContext.runAsUser="auto" \
+    --set securityContext.enabled=false \
+    --set shmVolume.chmod.enabled=false)
+fi
+
+echo "helm" "${args[@]}"
+helm "${args[@]}"
